fix the test

+++ storage/src/test/java/tech/pegasys/teku/storage/server/ChainStorageTest.java
@@ -306,9 +306,7 @@ public class ChainStorageTest {
       assertThatSafeFuture(blockResult).isCompletedWithEmptyOptional();
       final SafeFuture<List<SlotAndBlockRootAndBlobIndex>> sidecarKeysResult =
           chainStorage.getBlobSidecarKeys(
-              missingHistoricalBlock.getSlot(),
-              missingHistoricalBlock.getSlot(),
-              UInt64.valueOf(Long.MAX_VALUE));
+              missingHistoricalBlock.getSlot(), missingHistoricalBlock.getSlot(), Long.MAX_VALUE);
       assertThatSafeFuture(sidecarKeysResult).isCompletedWithValueMatching(List::isEmpty);
     }
